split projects.json into separate {project}.json files
