+++ apps/website/src/app/elf/database/projects/AmbrosiaMarketplace.json
@@ -1,0 +1,35 @@
+{
+    "$schema": "../schema/project.schema.json",
+    "ratingRaw": {
+        "impact": 4.2,
+        "difficulty": 4.0,
+        "size": 9.0
+    },
+    "title": "Ambrosia Marketplace",
+    "urls": {
+        "Github": {
+            "link": "https://github.com/amp7368/AmbrosiaMarketplace"
+        }
+    },
+    "dates": {
+        "dateApproximation": "DAY",
+        "startDate": "2022-03-13",
+        "endDate": "2022-07-01"
+    },
+    "summary": [
+        "A large project to develop a website for players to have storefronts for Wynncraft"
+    ],
+    "tools": {
+        "React": {},
+        "Typescript": {},
+        "@mui/material": {},
+        "SQL": {},
+        "NX": {},
+        "Docker": {},
+        "VSCode": {},
+        "TesseractJS": {}
+    },
+    "comments": [
+        "I'm learning how to train an AI to recognize Wynncraft font using TesseractJS."
+    ]
+}
