+++ apps/website/src/app/elf/database/projects/Voltskiya.json
@@ -1,0 +1,59 @@
+{
+    "$schema": "../schema/project.schema.json",
+    "title": "Voltskiya",
+    "ratingRaw": {
+        "impact": 8.0,
+        "difficulty": 7.5,
+        "size": 9.0
+    },
+    "urls": {
+        "Github-GameMechanics": {
+            "link": "https://github.com/amp7368/VoltskiyaGameMechanics",
+            "comment": "Repo for all major game mechanics"
+        },
+        "Github-CustomMobs": {
+            "link": "https://github.com/amp7368/VoltskiyaCustomMobs",
+            "comment": "Repo for all added custom mobs"
+        },
+        "Github-Utilities": {
+            "link": "https://github.com/amp7368/VoltskiyaUtilities",
+            "comment": "Repo for common utilities specific to Voltskiya"
+        },
+        "Github-Decay": {
+            "link": "https://github.com/amp7368/VoltskiyaDecay",
+            "comment": "Decay and regeneration of the map was big enough to be it's own repo"
+        }
+    },
+    "sections": [],
+    "dates": [
+        {
+            "dateApproximation": "MONTH",
+            "startDate": "2020-05-11",
+            "endDate": "2020-09-08"
+        },
+        {
+            "dateApproximation": "MONTH",
+            "startDate": "2021-03-01",
+            "endDate": "2021-10-30"
+        },
+        {
+            "dateApproximation": "MONTH",
+            "startDate": "2022-07-08",
+            "endDate": "2023-07-01"
+        }
+    ],
+    "summary": [
+        "Developed Minecraft Plugins to create a custom survival experience"
+    ],
+    "tools": {
+        "IntelliJ": {},
+        "Github": {},
+        "SQL": {},
+        "PaperMC": {},
+        "Java": {}
+    },
+    "comments": [
+        "I really grew as a programmer with this project more than any other to date.",
+        "Working to develop for an ever-changing environment taught me the importance of writing code that didn't just get the job done for today, but also for the next project"
+    ]
+}
