+++ apps/website/src/app/elf/database/projects/index.json
@@ -1,0 +1,10 @@
+[
+    "AmbrosiaCasino",
+    "AmbrosiaLoans",
+    "AmbrosiaMarketplace",
+    "appleptr16.com",
+    "BuildBot",
+    "Excursion",
+    "MAllyene",
+    "Voltskiya"
+]
